pyproject.toml
[project]
name = "ai-system-diagnoser"
version = "0.1.0"
description = "Local AI assistant for system diagnostics and optional fixes"
authors = [{name = "AI"}]
requires-python = ">=3.10"
dependencies = [
  "openai>=1.51.0",
  "psutil>=5.9.0",
  "python-dotenv>=1.0.1"
]

[project.optional-dependencies]
development = ["pytest>=7.4"]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
pythonpath = ["src"]
